. Alot of code cleanup. . (Reduced) Dumps should now be working . Triplify and Dump now use the same methods for accessing data
git-svn-id: http://linkedgeodata.googlecode.com/svn/trunk@74 00633da9-98c1-349d-95eb-89b68c819386

+++ src/main/java/org/linkedgeodata/jtriplify/methods/DefaultCoercions.java
@@ -103,10 +103,9 @@ public class DefaultCoercions
 	}
 
 
-
-	public static Integer toLong(Long a)
+	public static Long toLong(Integer a)
 	{
-		return a.intValue();
+		return a.longValue();
 	}
 	
 	public static Long toLong(Float a)
